+++ internal/apmconfig/env_test.go
@@ -14,23 +14,40 @@ func TestParseDurationEnv(t *testing.T) {
 	const envKey = "ELASTIC_APM_TEST_DURATION"
 	os.Setenv(envKey, "")
 
-	d, err := apmconfig.ParseDurationEnv(envKey, "s", 42*time.Second)
+	d, err := apmconfig.ParseDurationEnv(envKey, 42*time.Second)
 	assert.NoError(t, err)
 	assert.Equal(t, 42*time.Second, d)
 
-	os.Setenv(envKey, "5")
-	d, err = apmconfig.ParseDurationEnv(envKey, "s", 42*time.Second)
+	os.Setenv(envKey, "5s")
+	d, err = apmconfig.ParseDurationEnv(envKey, 42*time.Second)
 	assert.NoError(t, err)
 	assert.Equal(t, 5*time.Second, d)
 
 	os.Setenv(envKey, "5ms")
-	d, err = apmconfig.ParseDurationEnv(envKey, "s", 42*time.Second)
+	d, err = apmconfig.ParseDurationEnv(envKey, 42*time.Second)
 	assert.NoError(t, err)
 	assert.Equal(t, 5*time.Millisecond, d)
 
+	os.Setenv(envKey, "5m")
+	d, err = apmconfig.ParseDurationEnv(envKey, 42*time.Minute)
+	assert.NoError(t, err)
+	assert.Equal(t, 5*time.Minute, d)
+
+	os.Setenv(envKey, "5 h")
+	_, err = apmconfig.ParseDurationEnv(envKey, 42*time.Second)
+	assert.EqualError(t, err, "failed to parse ELASTIC_APM_TEST_DURATION: invalid character ' ' in duration 5 h")
+
+	os.Setenv(envKey, "5h")
+	_, err = apmconfig.ParseDurationEnv(envKey, 42*time.Second)
+	assert.EqualError(t, err, "failed to parse ELASTIC_APM_TEST_DURATION: invalid unit in duration 5h (allowed units: ms, s, m)")
+
 	os.Setenv(envKey, "5")
-	_, err = apmconfig.ParseDurationEnv(envKey, "", 42*time.Second)
-	assert.EqualError(t, err, "failed to parse ELASTIC_APM_TEST_DURATION: time: missing unit in duration 5")
+	_, err = apmconfig.ParseDurationEnv(envKey, 42*time.Second)
+	assert.EqualError(t, err, "failed to parse ELASTIC_APM_TEST_DURATION: missing unit in duration 5 (allowed units: ms, s, m)")
+
+	os.Setenv(envKey, "blah")
+	_, err = apmconfig.ParseDurationEnv(envKey, 42*time.Second)
+	assert.EqualError(t, err, "failed to parse ELASTIC_APM_TEST_DURATION: invalid duration blah")
 }
 
 func TestParseBoolEnv(t *testing.T) {